requirements.txt
# GroupMe bot dependencies
requests>=2.31.0
schedule>=1.2.0

# Web UI backend dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.10.0
